requirements.txt
pillow
requests
py7zr
webcolors
numpy
scikit-learn
colorthief
python-dotenv
